+++ src/main/java/com/ea/inzynierka/Application.java
@@ -1,12 +1,12 @@
-package com.ea.boot;
+package com.ea.inzynierka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@EnableJpaRepositories("com.ea.boot.repo")
-@EntityScan("com.ea.boot.model")
+@EnableJpaRepositories("com.ea.inzynierka.repo")
+@EntityScan("com.ea.inzynierka.model")
 @SpringBootApplication
 public class Application {
