+++ src/main/java/com/ea/inzynierka/configuration/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.ea.boot.configuration;
+package com.ea.inzynierka.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
