+++ src/main/java/com/ea/inzynierka/model/Book.java
@@ -1,4 +1,4 @@
-package com.ea.boot.model;
+package com.ea.inzynierka.model;
 
 import javax.persistence.*;
