+++ src/main/java/com/ea/inzynierka/web/SimpleController.java
@@ -1,4 +1,4 @@
-package com.ea.boot.web;
+package com.ea.inzynierka.web;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
