+++ src/main/resources/templates/list.html
@@ -37,7 +37,7 @@
             </table>
         </div>
         <div class="col-md-6">
-            <a class="btn btn-success" href="#" th:href="@{/books/create}"> Add Books </a>
+            <a class="btn btn-success" href="#" th:href="@{/books/add}"> Add Books </a>
 <!--
             <a class="btn btn-primary" href="#" th:href="@{/books/edit}" th:classappend="${bookList == null} ? 'disabled' : ''"> Edit Books </a>
 -->
